Move authorization to CanCan
This removes all of the previous functionality whereby a user was considered to
be an admin for a store because they existed in a stores' users, and not for
their roles.

Superadmin now refers to those users who can manage all aspects of the engine
and admin refers to users who can manage all aspects of a store (they must be in
that stores users).